[refactor] Split utils.py into separate files
